Adds f3SeqId to Container base class.
We need containers used for running F3 actions to mount their shared file
system volume, but a container/pod can't change attached volumes after being
created.  Since containers are reused, you could have a situation where two
sequences run the same action and are run in the same container - but the
volumes won't be correct.

By making f3SeqId part of the Container base class, we can use that when
selecting an appropriate container for an action.  If the action has an f3SeqId
parameter, that's used for selecting the container that the action is run in.